Refactor BotocoreStubber/BaseBackend into separate files
