+++ moto/core/__init__.py
@@ -1,4 +1,5 @@
-from .models import BaseBackend, get_account_id, ACCOUNT_ID  # noqa
+from .models import get_account_id, ACCOUNT_ID  # noqa
+from .base_backend import BaseBackend  # noqa
 from .common_models import BaseModel  # noqa
 from .common_models import CloudFormationModel, CloudWatchMetricProvider  # noqa
 from .models import patch_client, patch_resource  # noqa
