--- moto/core/urls.py
@@ -1,3 +1,0 @@
-url_bases = []
-
-url_paths = {}
