+++ moto/moto_api/_internal/models.py
@@ -1,4 +1,4 @@
-from moto.core.models import BaseBackend
+from moto.core import BaseBackend
 
 
 class MotoAPIBackend(BaseBackend):
